Refactor roles
